Move controllers to new route
